refactor: update Dify invocation configuration to use structured payload
- Changed the Dify invocation daemon to accept a structured payload for initialization, improving clarity and maintainability.
- Updated related configuration variables in the .env.example file to reflect the new naming convention for backwards invocation timeouts.
- Adjusted tests and plugin manager to accommodate the new payload structure.

+++ internal/core/dify_invocation/real/types.go
@@ -9,8 +9,8 @@ type RealBackwardsInvocation struct {
 	difyInnerApiKey     string
 	difyInnerApiBaseurl *url.URL
 	client              *http.Client
-	writeTimeout              int
-	readTimeout               int
+	writeTimeout        int64
+	readTimeout         int64
 }
 
 type BaseBackwardsInvocationResponse[T any] struct {
